requirements.txt
torch>=1.13
torchvision>=0.14
streamlit>=1.0
pandas>=1.0
opencv-python>=4.0
folium>=0.12
pydeck>=0.8
pillow>=8.0
pytest>=6.0
